pyproject.toml
[tool.poetry]
name = "counties_bot"
version = "0.1.0"
description = ""
authors = ["Denis Belavin"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
asyncpg = "^0.25.0"
aiogram = "^2.19"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
PyYAML = "^6.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
